dev(phpcs): check for php 5.6+ compat
