+++ .gitignore
@@ -1,2 +1,4 @@
 vendor
 composer.lock
+phpunit.xml
+phpcs.xml
