+++ composer.json
@@ -49,10 +49,12 @@
     "squizlabs/php_codesniffer": "^3.5",
     "brain/monkey": "^2.4",
     "phpunit/phpunit": "<= 9.3",
-    "mockery/mockery": "^1.3 | ^1.4"
+    "mockery/mockery": "^1.3 | ^1.4",
+    "dealerdirect/phpcodesniffer-composer-installer": "^0.7.0",
+    "phpcompatibility/php-compatibility": "^9.3"
   },
   "scripts": {
     "test": "vendor/bin/phpunit",
-    "lint": "vendor/bin/phpcs --standard=PSR12 --extensions=php --ignore=vendor/ -n -s src"
+    "lint": "vendor/bin/phpcs"
   }
 }
